refactor - class names, inject template name and template version
