+++ lib/highlander.factory.rb
@@ -1,8 +1,8 @@
 require_relative './highlander.dsl'
+require_relative './highlander.template.finder'
 require 'fileutils'
 require 'git'
 
-LOCAL_HIGHLANDER_CACHE_LOCATION = "#{ENV['HOME']}/.cfhighlander/components"
 
 module Highlander
 
@@ -26,14 +26,12 @@ module Highlander
           :cfndsl_ext_files,
           :lambda_src_files
 
-      def initialize(component_template_name,
-          component_template_version,
-          component_name, component_dir)
-        @template = component_template_name
+      def initialize(template_meta, component_name)
+        @template = template_meta
         @name = component_name
-        @component_dir = component_dir
+        @component_dir = template_meta.template_location
         @mappings = {}
-        @version = component_template_version
+        @version = template_meta.template_version
         @distribution_bucket = nil
         @distribution_prefix = nil
         @component_files = []
@@ -69,10 +67,10 @@ module Highlander
         end
 
 
-        @highlander_dsl_path = "#{@component_dir}/#{@template}.highlander.rb"
-        @cfndsl_path = "#{@component_dir}/#{@template}.cfndsl.rb"
+        @highlander_dsl_path = "#{@component_dir}/#{@template.template_name}.highlander.rb"
+        @cfndsl_path = "#{@component_dir}/#{@template.template_name}.cfndsl.rb"
         candidate_mappings_path = "#{@component_dir}/*.mappings.yaml"
-        candidate_dynamic_mappings_path = "#{@component_dir}/#{@template}.mappings.rb"
+        candidate_dynamic_mappings_path = "#{@component_dir}/#{@template.template_name}.mappings.rb"
 
         @cfndsl_ext_files += Dir["#{@component_dir}/ext/cfndsl/*.rb"]
         @lambda_src_files += Dir["#{@component_dir}/lambdas/**/*"].find_all {|p| not File.directory? p}
@@ -84,6 +82,8 @@ module Highlander
         @config.extend config_override unless config_override.nil?
         @config['component_version'] = @version unless @version.nil?
         @config['component_name'] = @name
+        @config['template_name'] = @template.template_name
+        @config['template_version'] = @template.template_version
         # allow override of components
         # config_override.each do |key, value|
         #   @config[key] = value
@@ -142,218 +142,25 @@ module Highlander
 
 
       def initialize(component_sources = [])
-        ## First look in local $PWD/components folder
-        ## Then search for cached $HOME/.highlander/components
-        ## Then search in sources given by dsl
-        default_locations = [
-            LOCAL_HIGHLANDER_CACHE_LOCATION,
-            File.expand_path('components'),
-            File.expand_path('.')
-        ]
-        default_locations.each do |predefined_path|
-          component_sources.unshift(predefined_path)
-        end
-
-        @component_sources = component_sources
-      end
-
-      def findTemplateDefault(template_name, component_version)
-        default_lookup_url = 'https://github.com/theonestack'
-        default_lookup_url = ENV['HIGHLANDER_DEFAULT_COMPONENT_GIT_LOOKUP'] if ENV.key? 'HIGHLANDER_DEFAULT_COMPONENT_GIT_LOOKUP'
-
-        git_url = "#{default_lookup_url}/hl-component-#{template_name}"
-
-        if component_version.nil? or component_version.empty? or component_version == 'latest'
-          branch = 'master'
-        else
-          branch = component_version
-        end
-        local_path = "#{LOCAL_HIGHLANDER_CACHE_LOCATION}/#{template_name}/#{component_version}"
-        return findTemplateGit(local_path, template_name, component_version, git_url, branch)
-
-      end
-
-      def findTemplateGit(cache_path, component_name, component_version, git_url, branch)
-        begin
-          cache_path = "#{cache_path}/" unless cache_path.end_with? '/'
-          # if this is snapshot, clean local cache
-          if branch.end_with? '.snapshot'
-            branch = branch.gsub('.snapshot', '')
-            FileUtils.rmtree cache_path if File.exist? cache_path and File.directory? cache_path
-          end
-
-          # if local cache exists, return from cache
-          if not Dir.glob("#{cache_path}*.highlander.rb").empty?
-            # if cache exists, just return from cache
-            component_name = Dir.glob("#{cache_path}*.highlander.rb")[0].gsub(cache_path, '').gsub('.highlander.rb', '')
-            return component_name, cache_path
-          end
-
-          # shallow clone
-          puts "Trying to load #{component_name}/#{component_version} from #{git_url}##{branch} ... "
-          clone_opts = { depth: 1 }
-          clone_opts[:branch] = branch if not (branch.nil? or branch.empty?)
-          Git.clone git_url, cache_path, clone_opts
-          puts "\t .. cached in #{cache_path}\n"
-          # return from cache once it's cloned
-          return findTemplateGit(cache_path, component_name, component_version, git_url, branch)
-        rescue Exception => e
-          STDERR.puts "Failed to resolve component #{component_name}@#{component_version} from #{git_url}"
-          STDERR.puts e
-          return nil
-        end
-      end
-
-      def findTemplateS3(s3_location, component_name, component_version)
-        parts = s3_location.split('/')
-        bucket = parts[2]
-        prefix = parts[3]
-        s3_key = "#{prefix}/#{component_name}/#{component_version}/#{component_name}.highlander.rb"
-        s3_prefix = "#{prefix}/#{component_name}/#{component_version}/"
-        local_destination = "#{LOCAL_HIGHLANDER_CACHE_LOCATION}/#{component_name}/#{component_version}"
-        begin
-          s3 = Aws::S3::Client.new({ region: s3_bucket_region(bucket) })
-          FileUtils.mkdir_p local_destination unless Dir.exist? local_destination
-
-          hl_content = s3.get_object({ bucket: bucket,
-              key: s3_key,
-              response_target: "#{local_destination}/#{component_name}.highlander.rb"
-          })
-          # if code execution got so far we consider file exists and download it locally
-          component_files = s3.list_objects_v2({ bucket: bucket, prefix: s3_prefix })
-          component_files.contents.each {|s3_object|
-            file_name = s3_object.key.gsub(s3_prefix, '')
-            destination_file = "#{local_destination}/#{file_name}"
-            destination_dir = File.dirname(destination_file)
-            print "Caching #{file_name} of #{component_name}@#{component_version} in #{destination_dir} ... "
-
-            FileUtils.mkpath(destination_dir) unless File.exists?(destination_dir)
-            s3.get_object({ bucket: bucket, key: s3_object.key, response_target: destination_file })
-            print " [OK] \n"
-          }
-          return local_destination
-        rescue => e
-          # this handles both nonexisting key and bucket
-          puts("#{component_name} not found in s3://#{bucket}/#{prefix}")
-          STDERR.puts(e.to_s) unless e.message.include? 'does not exist'
-          return nil
-        end
-      end
-
-      # will try and locate template based on git template location
-      # if template location is not git repo, returns nil
-      def tryFindTemplateGit(template_location, template_version)
-        if template_location.include? '#'
-          parts = template_location.split('#')
-          template_location = parts[0]
-          template_version = parts[1]
-        end
-
-        # avoid any nres
-        template_version = '' if template_version.nil?
-
-        # if empty or latest branch is empty
-        if template_version.empty? or template_version == 'latest' or template_version == 'latest.snapshot'
-          branch = ''
-        else
-          # otherwise component version is actual branch
-          branch = template_version
-        end
-
-
-        git_url = nil
-        if template_location.start_with? 'git:'
-          git_url = template_location.gsub('git:', '')
-        elsif template_location.start_with? 'github:'
-          git_url = "https://github.com/#{template_location.gsub('github:', '')}"
-        elsif template_location.start_with? 'github.com:'
-          git_url = "https://github.com/#{template_location.gsub('github.com:', '')}"
-        end
-
-        local_path = "#{LOCAL_HIGHLANDER_CACHE_LOCATION}/#{template_location}/#{template_version}"
-
-        if not git_url.nil?
-          template_name, location = findTemplateGit(local_path, template_location, template_version, git_url, branch)
-          if location.nil?
-            raise "Could not resolve component #{template_location}@#{template_version}"
-          else
-            return template_name, template_version, location
-          end
-        end
-
-        return nil
+        @template_finder = Highlander::Template::Finder.new(component_sources)
       end
 
       # Find component and given list of sources
       # @return [Highlander::Factory::Component]
       def loadComponentFromTemplate(template_name, template_version = nil, component_name = nil)
 
-        component_version_s = template_version.nil? ? 'latest' : template_version
-        template_version = nil if template_version == 'latest'
-        component_name = template_name if component_name.nil?
-
-        if template_name.include? '@' and (not template_name.start_with? 'git')
-          parts = template_name.split('@')
-          template_name = parts[0]
-          template_version = parts[1]
-          component_version_s = template_version
-        end
-
-        # if component specified as git location
-        new_template_name, new_version, candidate_git = tryFindTemplateGit(template_name, component_version_s)
-        return buildComponentFromLocation(new_template_name, new_version, component_name, candidate_git) unless candidate_git.nil?
-
-        # if not git but has .snapshot lookup in default
-        if (not template_version.nil?) and template_version.end_with? '.snapshot'
-          new_template_name, default_candidate = findTemplateDefault(template_name, component_version_s)
-          return buildComponentFromLocation(new_template_name, template_version, component_name, default_candidate) unless default_candidate.nil?
-        end
-
-        # try in all of the component source
-        @component_sources.each do |source|
-          component_full_name = "#{template_name}@#{template_version.nil? ? 'latest' : template_version}"
-          # TODO handle http(s) sources and their download to local
-          if source.start_with?('http')
-            raise StandardError, 'http(s) sources not supported yet'
-          elsif source.start_with?('s3://')
-            # s3 candidate
-
-            s3_candidate = findTemplateS3(source, template_name, component_version_s)
-            if not s3_candidate.nil?
-              # at this point all component files are download to local file system and consumed from there
-              return buildComponentFromLocation(template_name, template_version, component_name, s3_candidate)
-            end
-
-          else
-            # file system candidate
-            candidate = "#{source}/#{template_name}"
-            candidate = "#{candidate}/#{template_version}" unless template_version.nil?
-            candidate_hl_path = "#{candidate}/#{template_name}.highlander.rb"
-            candidate2_hl_path = "#{source}/#{template_name}.highlander.rb"
-            puts "Trying to load #{component_full_name} from #{candidate} ... "
-            if File.exist?(candidate_hl_path)
-              return buildComponentFromLocation(template_name, template_version, component_name, candidate)
-            end
-            puts "Trying to load #{component_full_name} from #{source} ... "
-            if File.exist?(candidate2_hl_path)
-              return buildComponentFromLocation(template_name, template_version, component_name, source)
-            end unless component_version_s != 'latest'
-          end
-        end
-
-        # try default component source on github
-        template_name, default_candidate = findTemplateDefault(template_name, component_version_s)
-        return buildComponentFromLocation(template_name, template_version, component_name, default_candidate) unless default_candidate.nil?
+        template_meta = @template_finder.findTemplate(template_name, template_version)
 
         raise StandardError, "highlander template #{template_name}@#{component_version_s} not located" +
-            " in sources #{@component_sources}"
+            " in sources #{@component_sources}" if template_meta.nil?
+
+        return buildComponentFromLocation(template_meta, component_name)
+
       end
 
-      def buildComponentFromLocation(template_name,
-          template_version,
-          component_name,
-          template_location)
-        component = Component.new(template_name, template_version,  component_name, template_location)
+
+      def buildComponentFromLocation(template_meta, component_name)
+        component = Component.new(template_meta, component_name)
         component.load_config
         return component
       end
